Changed 'uint32_t' to 'size_t'.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@102097 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VBoxManage/VBoxManageCloud.cpp
@@ -1293,7 +1293,7 @@ static RTEXITCODE cloudInstanceMetricList(HandlerArg *a, int iFirst, PCLOUDCOMMO
                      RTEXITCODE_FAILURE);
 
     RTPrintf(Cloud::tr("Available metric names:\n"));
-    uint32_t cMetricNamesArraySize = metricNamesArray.size();
+    size_t cMetricNamesArraySize = metricNamesArray.size();
 
     if (cMetricNamesArraySize == 0)
     {
@@ -1303,7 +1303,7 @@ static RTEXITCODE cloudInstanceMetricList(HandlerArg *a, int iFirst, PCLOUDCOMMO
     else
     {
         Bstr value;
-        for (uint32_t k = 0; k < cMetricNamesArraySize; ++k)
+        for (size_t k = 0; k < cMetricNamesArraySize; ++k)
         {
             value = metricNamesArray[k];
             MetricType_T aMetricType;
